+++ rfront/src/Components/Home/Section1.js
@@ -15,73 +15,13 @@ import Grid from '@mui/material/Grid';
 import { Box } from '@mui/system';
 
 
-
-function card_cont(imagen, titulo, descripcion, btn) {
-
-    return (
-        <div>
-            <Card
-                style={{ backgroundColor: "#ffffff" }}
-                sx={{ alignItems: 'center', height: '100%', borderRadius: '20px', display: 'flex', flexDirection: 'column', minWidth: 225, maxWidth: 400, minHeight: 220, maxHeight: 350, paddingTop: '20px' }}
-                md={6}
-            >
-                <CardMedia
-
-                    component="img"
-                    sx={{
-                        height: 120,
-                        width: 120,
-                    }}
-                    md={6}
-                    image={imagen}
-                    alt="random" />
-                <CardContent sx={{ flexGrow: 1 }}>
-                    <Typography align='center' gutterBottom variant="h5" component="h2">
-                        {titulo}
-                    </Typography>
-                    <Typography align='center'>
-                        {descripcion}
-                    </Typography>
-                </CardContent>
-            </Card>
-        </div>
-    );
-}
-
-
-
 export default function Section1(){
     
     
     return (
-        <Grid container className='seccion-lideres'>
-            
-                <Grid sx={6} sm={6} md={6}>
-                    <Stack alignItems='flex-end' margin='3%'>
-                    {card_cont(pets, 'Mascotas', ' En este sitio web encontraras una lista de negocios cercanos a tu ubicación, y los mejores productos y servicios para tus mascotas.')}
-                    </Stack>
-                </Grid>
-                <Grid sx={6} sm={6} md={6} >
-                    <Stack alignItems='flex-start' margin='3%'>
-                    {card_cont(empresas, 'Empresas', 'Somos un directorio especializado en productos y servicios para mascotas. Publíca tu negocio y capta nuevos clientes en tu localidad.')}
-                    </Stack>
-                </Grid>
-        
-
-            <div>
-                <Typography variant="h4" align="center" paddingTop="50px" component="div" mb='0px'>
-                    Para Líderes de Manada
-                </Typography>
-                <Typography variant="h5" align="center" mt='-20px'>
-                    ________________________
-                </Typography>
-                <Typography variant="h6" align="center" component="div" mb='20px'>
-                    ¿Cómo funciona?
-                </Typography>
-
-                <Stack className='lideres' direction='row' justifyContent='center' spacing={4} marginX='100px'>
-
-                    <Stack>
+        <Grid container alignItems='center' justifyContent='center' className='seccion-lideres'>
+      
+                    <Grid sx={12} sm={12} md={6}>
                         <Typography fontSize='24' align="left" component="div" mb='0px'>
                             Encontrar proveedores cercanos
                         </Typography>
@@ -116,15 +56,12 @@ export default function Section1(){
                             </Button>
                         </div>
 
-                    </Stack>
+                    </Grid>
 
-                    <Stack>
+                    <Grid sx={12} sm={12} md={6} justifyContent='center' className='imagen-people-pet'>
                         <img className='img-people-pet' src={peopleandpet} alt="peopleandpet" />
-                    </Stack>
+                    </Grid>
 
-                </Stack>
-
-            </div>
         </Grid>
     )
